[🥁 : feat] 타임라인 데이터 패치 과정 커스텀 훅으로 분리
- 커스텀 훅 레포 최초 생성에 따른 절대 경로 지정
- 'tweets' 컬렉션에서 데이터를 가져오는 과정이 반복적으로 일어나며 차이점은 where 쿼리 조건. where문은 각 컴포넌트에서 작성해 배열로 전달해주는 형식을 사용하며 타입은 any 활용
- orderBy('createdAt', desc)가 조건에 있음에도 최종적으로 sort를 한 번 더 진행하는 이유는 search-timeline 때문. search-timeline에서는 where문을 위해 무조건 orderBy('tweet')을 우선 작성해주어야했기 때문. 검색 로직을 변경하게된다면 이 부분은 삭제 가능할 것으로 예상함
