+++ src/hooks/useTimeline.tsx
@@ -1,0 +1,49 @@
+import { useEffect, useState } from 'react';
+import { Unsubscribe } from 'firebase/auth';
+import {
+  collection,
+  limit,
+  onSnapshot,
+  orderBy,
+  query,
+} from 'firebase/firestore';
+import { db } from '@/firebase.ts';
+import ITweet from '@type/ITweet.ts';
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export default function useTimeline(queryOptions: any[]) {
+  const [tweets, setTweets] = useState<ITweet[]>([]);
+  useEffect(() => {
+    let unsubscribe: Unsubscribe | null = null;
+    const fetchTweets = async () => {
+      const tweetsQuery = query(
+        collection(db, 'tweets'),
+        ...queryOptions,
+        orderBy('createdAt', 'desc'),
+        limit(25),
+      );
+      unsubscribe = await onSnapshot(tweetsQuery, (snapshot) => {
+        const tweetList = snapshot.docs.map((doc) => {
+          const { userId, userName, tweet, createdAt, photo } = doc.data();
+          return {
+            userId,
+            userName,
+            tweet,
+            createdAt,
+            photo,
+            id: doc.id,
+          };
+        });
+        tweetList.sort((a, b) => b.createdAt - a.createdAt);
+        setTweets(tweetList);
+      });
+    };
+    fetchTweets();
+    return () => {
+      if (unsubscribe) {
+        unsubscribe();
+      }
+    };
+  }, [queryOptions]);
+  return tweets;
+}
